Update: COnfiguracion centralizada, mejoras en la interfaz y readme añadido. Ademas, alguno popups de ayuda para instalar la aplicacion web facilmente, readme actualizado tambien.

+++ delete.php
@@ -1,41 +1,54 @@
 <?php
-
-$config = include 'config.php';
+include 'main.php';
 
 $resultado = [
   'error' => false,
   'mensaje' => ''
 ];
 
-if(isset($_POST['submit_delete'])){
+   if (isset($_POST['submit_delete']) && $_POST['id'] != '') {
 
-  try {
-  $dsn = 'mysql:host=' . $config['db']['host'] . ';dbname=' . $config['db']['name'];
-  $conexion = new PDO($dsn, $config['db']['user'], $config['db']['pass'], $config['db']['options']);
-    
-  $id = $_POST['id'];
-  $consultaSQL = "DELETE FROM empleados WHERE id =" . $id;
+      $conexion = connection("try");
 
-  $delete = $conexion->prepare($consultaSQL);
-  $delete->execute();
+      if($conexion['status']){
+         $conexion = connection("con");
 
-  $resultado = [
-         'error' => false,
-         'mensaje' => 'El  registro fue eliminado correctamente.'
-   ];
+         $id = $_POST['id'];
+         $consultaSQL = "DELETE FROM empleados WHERE id =" . $id;
 
-} catch(PDOException $error) {
-  $resultado['error'] = true;
-  $resultado['mensaje'] = $error->getMessage();
-}
+         $delete = $conexion->prepare($consultaSQL);
+         
+         try{
+            if($delete->execute()){
+               $resultado = [
+                  'error' => false,
+                  'mensaje' => 'El  registro fue eliminado correctamente.'
+               ];
+            }
+         } catch(Exception $error){
+            $resultado = [
+               'error' => true,
+               'mensaje' => $error->getMessage()
+            ];
+         }
 
-} else{
-  $resultado = [
-    'error' => true,
-    'mensaje' => '¡Hubo un error! El formulario no se envio correctamente.'
-   ];
-}
+         /*
+         if($delete->execute()){
+            $resultado = [
+            'error' => false,
+            'mensaje' => 'El  registro fue eliminado correctamente.'
+            ];
+         }
+         */
 
+      } else{
+         $resultado = [
+            'error' => true,
+            'mensaje' => $conexion['error']
+         ];
+      }
+      
+   }
 ?>
 
 <!DOCTYPE html>
@@ -46,6 +59,25 @@ if(isset($_POST['submit_delete'])){
     <meta http-equiv="x-ua-compatible" content="ie=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
 
+    <!---     | Favicon |     -->
+   <link rel="apple-touch-icon" sizes="57x57" href="IMG/favicon/apple-icon-57x57.png">
+   <link rel="apple-touch-icon" sizes="60x60" href="IMG/favicon/apple-icon-60x60.png">
+   <link rel="apple-touch-icon" sizes="72x72" href="IMG/favicon/apple-icon-72x72.png">
+   <link rel="apple-touch-icon" sizes="76x76" href="IMG/favicon/apple-icon-76x76.png">
+   <link rel="apple-touch-icon" sizes="114x114" href="IMG/favicon/apple-icon-114x114.png">
+   <link rel="apple-touch-icon" sizes="120x120" href="IMG/favicon/apple-icon-120x120.png">
+   <link rel="apple-touch-icon" sizes="144x144" href="IMG/favicon/apple-icon-144x144.png">
+   <link rel="apple-touch-icon" sizes="152x152" href="IMG/favicon/apple-icon-152x152.png">
+   <link rel="apple-touch-icon" sizes="180x180" href="IMG/favicon/apple-icon-180x180.png">
+   <link rel="icon" type="image/png" sizes="192x192"  href="IMG/favicon/android-icon-192x192.png">
+   <link rel="icon" type="image/png" sizes="32x32" href="IMG/favicon/favicon-32x32.png">
+   <link rel="icon" type="image/png" sizes="96x96" href="IMG/favicon/favicon-96x96.png">
+   <link rel="icon" type="image/png" sizes="16x16" href="IMG/favicon/favicon-16x16.png">
+   <link rel="manifest" href="/manifest.json">
+   <meta name="msapplication-TileColor" content="#ffffff">
+   <meta name="msapplication-TileImage" content="/ms-icon-144x144.png">
+   <meta name="theme-color" content="#ffffff">
+
     <!---     | CSS |     -->
     <link rel="stylesheet" href="CSS/w3.css"/>
     <link rel="stylesheet" href="CSS/font_awesome.css">
@@ -54,7 +86,17 @@ if(isset($_POST['submit_delete'])){
     <script  type="text/javascript" src="JS/main.js"></script>
 
    <!---     | Titulo |     -->
-   <title>Usuario eliminado</title>
+   <?php
+   if (isset($resultado)) {
+
+      if (!$resultado['error']){
+         echo '<title>Registro eliminado</title>'; 
+      }else{
+         echo '<title>Error durante el proceso</title>';
+      }
+
+   }
+   ?>
 
 
 </head>
@@ -62,61 +104,12 @@ if(isset($_POST['submit_delete'])){
      
 <?php
 
-// Proceso exitoso, popup.
-   $success = '
-      <div id="popup" class="w3-modal">
-      <div class="w3-modal-content w3-animate-zoom">
+if(isset($resultado)) {
 
-         <header class="w3-container w3-blue w3-center">
-            <h2> Proceso terminado con exito!</h2>
-         </header>
-
-         <div class="w3-container w3-center">
-
-             <img src="IMG/Success.png" style="width:20%" style="width:30%" class="w3-circle w3-margin-top">
-
-            <p>'.$resultado['mensaje'].'</p>
-            <p>Por favor, preciona "volver" para volver a la pagina principal.</p>
-
-            <a href="index.php" class="w3-button w3-round w3-blue w3-margin-bottom"><i class="fa fa-arrow-left"></i> Volver</a>
-         </div>
-
-      </div>
-   </div>
-
-   ';
-
-   // Error en el proceso, popup.
-   $error = '
-      <div id="popup" class="w3-modal">
-      <div class="w3-modal-content w3-animate-zoom">
-
-         <header class="w3-container w3-red w3-center">
-            <h2> Hubo un error!</h2>
-         </header>
-
-         <div class="w3-container w3-center">
-
-            <img src="IMG/Error.png" style="width:20%" style="width:30%" class="w3-circle w3-margin-top">
-
-            <p>'.$resultado['mensaje'].'</p>
-            <p>Por favor preciona "volver" para intentarlo nuevamente.</p>
-            <p>Si el error perciste consulte con el administrador de la pagina.</p>
-
-            <a href="index.php" class="w3-button w3-round w3-red w3-margin-bottom"><i class="fa fa-arrow-left"></i> Volver</a>
-         </div>
-
-      </div>
-   </div>
-
-   ';
-
-if (isset($resultado)) {
-
-   if (!$resultado['error']){
-        echo $success; 
+   if(!$resultado['error']){
+        displaymessage('success',$resultado['mensaje']); 
    }else{
-      echo $error;
+      displaymessage('error',$resultado['mensaje']);
    }
 
 }
